Overhaul monolithic frontend into react components
